services/api/app/services/drive.py
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
from app.core.config import settings

class DriveService:
    def __init__(self, token: str):
        self.creds = Credentials(token)
        self.service = build('drive', 'v3', credentials=self.creds)

    def list_files(self, query: str = None, page_size: int = 10):
        results = self.service.files().list(
            q=query, pageSize=page_size, fields="nextPageToken, files(id, name, mimeType)").execute()
        return results.get('files', [])

    def download_file(self, file_id: str):
        # Implementation to download file content for RAG
        pass
    
    def watch_changes(self, channel_id: str):
        # Implementation to set up push notifications for file changes
        pass
